Fix intermittent "Could not start main server"
The symptom is the error:

   EDTS: Could not start main server

If there is I/O while evaluating sit-for, the sit-for gets
interrupted, for Emacs to process I/O.  So if there was I/O when edts
was starting a node, the retry loop in edts-api-start-server could
exhaust in almost no time, before the node has started.  For me this
happened more often when starting emacs with a file like: emacs
some_file.erl, but it did not happen every time.

Instead of retrying a certain fixed number of times, retry until a
computed end-time has passed.