Don't initialize xref concurrently
It has been observed that sometimes the project node starts to consume
100% cpu indefinitely, and it seems to be related to the xref_server.

For a large project, such as opening a file in Erlang/OTP, the
edts_xref_server:update_paths/2 could take a significant amount of
time.

The edts_xref_server:update_paths/2 loads project modules, then
refreshes these to the xref_server, then unloads modules.
If the user opens another file during this process, a second
process will start to do loading and unloading in parallel, and
the project node will sometimes start to consume 100% cpu.

This commit uses global:trans to make sure the code of start/0 is not
executed concurrently. A second call to this will now block waiting
for the first to complete.  The Emacs already runs this in the
background, so it will still be responsive, only some xref services
will not immediately be available until the xref initialization has
completed.