Load code and start services only once per node
Edts initializes the node (or a new node) every time a new erlang file
is opened.  Initializing a new node involves loading edts onto the
node, but if the user twice quickly file while a long-lived background
process is running, it will get killed by the code loading mechanism.
An example when this can happen is when the xref plugin initializes
for a large project.

The strategy to always initialize the node was discussed in
https://github.com/tjarvstrand/edts/issues/237 and one motivation
to do node_init always was "to update code path, xref etc if you open
a file in a newly created directory (eg. app)", see the comment
https://github.com/tjarvstrand/edts/issues/237#issuecomment-373703531

This commit attempts to not break code path updates, while
avoid errors caused by the code loading.